requirements.txt
fastapi
httpx
firebase-admin
spacy
python-dotenv
sendgrid
websockets
PyMuPDF
textwrap3
chromadb
sentence-transformers
tiktoken
